Move `Simulation` inside `Circuit`.
