+++ tket/src/Circuit/Simulation/BitOperations.cpp
