+++ tket/src/Circuit/Simulation/BitOperations.hpp
