+++ tket/src/Circuit/Simulation/GateNode.cpp
