+++ tket/src/Circuit/Simulation/GateNode.hpp
