+++ tket/src/Circuit/Simulation/GateNodesBuffer.cpp
