+++ tket/src/Circuit/Simulation/GateNodesBuffer.hpp
