+++ tket/src/Circuit/Simulation/PauliExpBoxUnitaryCalculator.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "tket/Simulation/PauliExpBoxUnitaryCalculator.hpp"
+#include "tket/Circuit/Simulation/PauliExpBoxUnitaryCalculator.hpp"
 
 #include <algorithm>
 #include <tkassert/Assert.hpp>
