+++ tket/test/src/Circuit/test_Boxes.cpp
@@ -21,9 +21,9 @@
 #include "tket/Circuit/CircUtils.hpp"
 #include "tket/Circuit/Circuit.hpp"
 #include "tket/Circuit/PauliExpBoxes.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Converters/PhasePoly.hpp"
 #include "tket/Gate/SymTable.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 
 using Catch::Matchers::StartsWith;
