+++ tket/test/src/Circuit/test_Circ.cpp
@@ -23,6 +23,7 @@
 #include "tket/Circuit/CircUtils.hpp"
 #include "tket/Circuit/Circuit.hpp"
 #include "tket/Circuit/DAGDefs.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Gate/GatePtr.hpp"
 #include "tket/Gate/OpPtrFunctions.hpp"
 #include "tket/OpType/EdgeType.hpp"
@@ -31,7 +32,6 @@
 #include "tket/Ops/ClassicalOps.hpp"
 #include "tket/Ops/Op.hpp"
 #include "tket/Ops/OpPtr.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 #include "tket/Transformations/Decomposition.hpp"
 #include "tket/Transformations/OptimisationPass.hpp"
 #include "tket/Transformations/Replacement.hpp"
