+++ tket/test/src/Simulation/test_CircuitSimulator.cpp
@@ -24,9 +24,9 @@
 #include "ComparisonFunctions.hpp"
 #include "tket/Circuit/CircPool.hpp"
 #include "tket/Circuit/CircUtils.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Gate/GateUnitaryMatrix.hpp"
 #include "tket/Gate/GateUnitaryMatrixUtils.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 #include "tket/Transformations/OptimisationPass.hpp"
 #include "tket/Transformations/Transform.hpp"
 #include "tket/Utils/MatrixAnalysis.hpp"
