+++ tket/test/src/Simulation/test_PauliExpBoxUnitaryCalculator.cpp
@@ -17,8 +17,8 @@
 
 #include "tket/Circuit/Boxes.hpp"
 #include "tket/Circuit/PauliExpBoxes.hpp"
+#include "tket/Circuit/Simulation/PauliExpBoxUnitaryCalculator.hpp"
 #include "tket/Gate/GateUnitaryMatrix.hpp"
-#include "tket/Simulation/PauliExpBoxUnitaryCalculator.hpp"
 
 // This is for testing PauliExpBoxUnitaryCalculator, an internal
 // component of the Simulation code.
