+++ tket/test/src/test_ArchitectureAwareSynthesis.cpp
@@ -17,9 +17,9 @@
 #include "Simulation/ComparisonFunctions.hpp"
 #include "testutil.hpp"
 #include "tket/Architecture/Architecture.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Predicates/CompilerPass.hpp"
 #include "tket/Predicates/PassGenerators.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 
 namespace tket {
 using Connection = Architecture::Connection;
