+++ tket/test/src/test_Clifford.cpp
@@ -19,8 +19,8 @@
 #include "Simulation/ComparisonFunctions.hpp"
 #include "testutil.hpp"
 #include "tket/Circuit/CircUtils.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Ops/ClassicalOps.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 #include "tket/Transformations/CliffordOptimisation.hpp"
 #include "tket/Transformations/CliffordReductionPass.hpp"
 #include "tket/Transformations/Decomposition.hpp"
