+++ tket/test/src/test_ContextOpt.cpp
@@ -16,10 +16,10 @@
 
 #include "Simulation/ComparisonFunctions.hpp"
 #include "tket/Circuit/Circuit.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Predicates/CompilationUnit.hpp"
 #include "tket/Predicates/CompilerPass.hpp"
 #include "tket/Predicates/PassGenerators.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 #include "tket/Transformations/ContextualReduction.hpp"
 #include "tket/Transformations/Transform.hpp"
 #include "tket/Utils/EigenConfig.hpp"
