+++ tket/test/src/test_PhasedXFrontier.cpp
@@ -17,7 +17,7 @@
 
 #include "testutil.hpp"
 #include "tket/Circuit/Circuit.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Transformations/Decomposition.hpp"
 #include "tket/Transformations/PhasedXFrontier.hpp"
 #include "tket/Utils/Expression.hpp"
