+++ tket/test/src/test_TwoQubitCanonical.cpp
@@ -18,11 +18,11 @@
 #include "testutil.hpp"
 #include "tket/Circuit/CircUtils.hpp"
 #include "tket/Circuit/Command.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Gate/Rotation.hpp"
 #include "tket/Ops/ClassicalOps.hpp"
 #include "tket/Predicates/CompilationUnit.hpp"
 #include "tket/Predicates/PassGenerators.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
 #include "tket/Transformations/BasicOptimisation.hpp"
 #include "tket/Transformations/Decomposition.hpp"
 #include "tket/Transformations/Transform.hpp"
