+++ tket/test/src/testutil.cpp
@@ -18,7 +18,7 @@
 
 #include "Simulation/ComparisonFunctions.hpp"
 #include "tket/Circuit/Circuit.hpp"
-#include "tket/Simulation/CircuitSimulator.hpp"
+#include "tket/Circuit/Simulation/CircuitSimulator.hpp"
 #include "tket/Utils/EigenConfig.hpp"
 #include "tket/Utils/MatrixAnalysis.hpp"
